+++ dandy/core/url.py
@@ -10,8 +10,8 @@ class Url:
     path_parameters: List[str] = field(default_factory=list)
     query_parameters: Dict[str, str] = field(default_factory=dict)
 
-    def __str__(self):
-        self.to_str()
+    def __str__(self) -> str:
+        return self.to_str()
 
     @property
     def parsed_url(self) -> ParseResult: