+++ dandy/future/future.py
@@ -4,7 +4,7 @@ from time import time, perf_counter
 
 from typing_extensions import Callable, TypeVar, Generic, Union
 
-from dandy.future.exceptions import FutureException
+from dandy.future.exceptions import FutureCriticalException
 
 async_executor = concurrent.futures.ThreadPoolExecutor()
 
@@ -39,7 +39,7 @@ class AsyncFuture(Generic[FutureResultType]):
             else:
                 raise concurrent.futures.TimeoutError
         except concurrent.futures.TimeoutError:
-            raise FutureException(f'Future timed out after {self._result_timeout} seconds')
+            raise FutureCriticalException(f'Future timed out after {self._result_timeout} seconds')
         finally:
             del self._future